Правки по улучшению кода

+++ homework_3/src/main/java/ru/parfenov/homework_3/servlets/admin/DeleteUserServlet.java
@@ -30,13 +30,10 @@ public class DeleteUserServlet extends HttpServlet {
     @Override
     public void doDelete(HttpServletRequest request, HttpServletResponse response) throws IOException {
         HttpSession session = request.getSession();
-        int responseStatus;
         var user = (User) session.getAttribute("user");
-        String jsonString;
-        if (user == null || user.getRole() != UserRole.ADMIN) {
-            jsonString = "no rights or registration!";
-            responseStatus = user == null ? 401 : 403;
-        } else {
+        int responseStatus = user == null ? 401 : 403;
+        String jsonString = "no rights or registration!";
+        if (user != null && user.getRole() == UserRole.ADMIN) {
             String userIdStr = request.getParameter("id");
             jsonString = userService.delete(userIdStr) ? "the user is deleted" : "the user is not deleted!";
             responseStatus = "the user is not deleted!".equals(jsonString) ? 404 : 200;
